flatten components, remove redundant files
